+++ CHANGELOG.rst
@@ -1,6 +1,18 @@
 Changelog
 #########
 
+2.0.0
+*****
+
+- added ``$canonicalizeKeys`` parameter to:
+
+  - ``ComparisonHelper::isLooselyIdentical()``
+  - ``AssertionTrait::assertLooselyIdentical()``
+  - ``AssertionTrait::assertLooselyIdenticalIterable()``
+  - ``AssertionTrait::looselyIdenticalTo()``
+  - ``AssertionTrait::looselyIdenticalIterable()``
+
+
 1.0.0
 *****